Don't re-download the download cache for sanity reasons

+++ short/run-v3.sh
@@ -7,9 +7,11 @@ source $SCRIPT_DIR/../set_test_env.sh
 rm -rf $HOME/.phoronix-test-suite/installed-tests/
 
 # get the download cache for the short test suite
-wget https://people.canonical.com/~mckeesh/short-dl-cache.tar.xz
-tar xvf short-dl-cache.tar.xz
-mv download-cache $HOME/.phoronix-test-suite
+if [ ! -f short-dl-cache.tar.xz ]; then
+    wget https://people.canonical.com/~mckeesh/short-dl-cache.tar.xz
+    tar xvf short-dl-cache.tar.xz
+    mv download-cache $HOME/.phoronix-test-suite
+fi
 
 export TEST_RESULTS_IDENTIFIER=short-v3-id 
 export TEST_RESULTS_NAME=short-v3-name
